requirements.txt
numpy==1.24.3
streamlit==1.34.0
tensorflow==2.14.0
Pillow==9.2.0

